+++ src/internal-links/helpers.js
@@ -10,7 +10,6 @@
  * governing permissions and limitations under the License.
  */
 import { tracingFetch as fetch } from '@adobe/spacecat-shared-utils';
-import { AbortController, AbortError } from '@adobe/fetch';
 
 const LINK_TIMEOUT = 3000;
 export const CPC_DEFAULT_VALUE = 1;
@@ -78,12 +77,8 @@ export const calculateKpiDeltasForAudit = (auditData) => {
  * false if reachable/accessible
  */
 export async function isLinkInaccessible(url, log) {
-  const controller = new AbortController();
-  const timeoutId = setTimeout(() => controller.abort(), LINK_TIMEOUT);
-
   try {
-    const response = await fetch(url, { signal: controller.signal });
-    clearTimeout(timeoutId);
+    const response = await fetch(url, { timeout: LINK_TIMEOUT });
     const { status } = response;
 
     // Log non-404, non-200 status codes
@@ -94,8 +89,7 @@ export async function isLinkInaccessible(url, log) {
     // URL is valid if status code is less than 400, otherwise it is invalid
     return status >= 400;
   } catch (error) {
-    clearTimeout(timeoutId);
-    log.info(`broken-internal-links audit: Error checking ${url}: ${error instanceof AbortError ? `Request timed out after ${LINK_TIMEOUT}ms` : error.message}`);
+    log.info(`broken-internal-links audit: Error checking ${url}: ${error.code === 'ETIMEOUT' ? `Request timed out after ${LINK_TIMEOUT}ms` : error.message}`);
     // Any error means the URL is inaccessible
     return true;
   }