+++ src/metatags/handler.js
@@ -16,12 +16,17 @@ import { calculateCPCValue } from '../support/utils.js';
 import { getObjectFromKey, getObjectKeysUsingPrefix } from '../utils/s3-utils.js';
 import SeoChecks from './seo-checks.js';
 import { AuditBuilder } from '../common/audit-builder.js';
-import { noopUrlResolver, wwwUrlResolver } from '../common/index.js';
+import { wwwUrlResolver } from '../common/index.js';
 import metatagsAutoSuggest from './metatags-auto-suggest.js';
 import { convertToOpportunity } from '../common/opportunity.js';
 import { getTopPagesForSiteId } from '../canonical/handler.js';
 import { getIssueRanking, removeTrailingSlash } from './opportunity-utils.js';
-import { DESCRIPTION, H1, TITLE } from './constants.js';
+import {
+  DESCRIPTION,
+  H1,
+  PROJECTED_VALUE_THRESHOLD,
+  TITLE,
+} from './constants.js';
 import { syncSuggestions } from '../utils/data-access.js';
 import { createOpportunityData } from './opportunity-data-mapper.js';
 
@@ -137,42 +142,56 @@ function getOrganicTrafficForEndpoint(endpoint, rumDataMapMonthly, rumDataMapBiM
 }
 
 // Calculate the projected traffic lost for a site
-async function calculateProjectedTraffic(context, site, detectedTags, log) {
-  const rumAPIClient = RUMAPIClient.createFrom(context);
+async function calculateProjectedTraffic(context, auditUrl, siteId, detectedTags, log) {
   const options = {
-    domain: wwwUrlResolver(site),
+    domain: auditUrl,
     interval: 30,
     granularity: 'DAILY',
   };
-  const queryResultsMonthly = await rumAPIClient.query('traffic-acquisition', options);
-  const queryResultsBiMonthly = await rumAPIClient.query('traffic-acquisition', {
-    ...options,
-    interval: 60,
-  });
-  const { rumDataMapMonthly, rumDataMapBiMonthly } = preprocessRumData(
-    queryResultsMonthly,
-    queryResultsBiMonthly,
-  );
-  let projectedTraffic = 0;
-  Object.entries(detectedTags).forEach(([endpoint, tags]) => {
-    const organicTraffic = getOrganicTrafficForEndpoint(
-      endpoint,
-      rumDataMapMonthly,
-      rumDataMapBiMonthly,
-      log,
-    );
-    Object.values((tags)).forEach((tagIssueDetails) => {
-      // Multiplying by 1% for missing tags, and 0.5% for other tag issues
-      // For duplicate tags, each page's traffic is multiplied by .5% so
-      // it amounts to 0.5% * number of duplicates.
-      const multiplier = tagIssueDetails.issue.includes('Missing') ? 0.01 : 0.005;
-      projectedTraffic += organicTraffic * multiplier;
+  try {
+    const rumAPIClient = RUMAPIClient.createFrom(context);
+    const queryResultsMonthly = await rumAPIClient.query('traffic-acquisition', options);
+    const queryResultsBiMonthly = await rumAPIClient.query('traffic-acquisition', {
+      ...options,
+      interval: 60,
     });
-  });
-  return projectedTraffic;
+
+    const { rumDataMapMonthly, rumDataMapBiMonthly } = preprocessRumData(
+      queryResultsMonthly,
+      queryResultsBiMonthly,
+    );
+
+    let projectedTrafficLost = 0;
+    Object.entries(detectedTags).forEach(([endpoint, tags]) => {
+      const organicTraffic = getOrganicTrafficForEndpoint(
+        endpoint,
+        rumDataMapMonthly,
+        rumDataMapBiMonthly,
+        log,
+      );
+      Object.values((tags)).forEach((tagIssueDetails) => {
+        // Multiplying by 1% for missing tags, and 0.5% for other tag issues
+        // For duplicate tags, each page's traffic is multiplied by .5% so
+        // it amounts to 0.5% * number of duplicates.
+        const multiplier = tagIssueDetails.issue.includes('Missing') ? 0.01 : 0.005;
+        projectedTrafficLost += organicTraffic * multiplier;
+      });
+    });
+
+    const cpcValue = await calculateCPCValue(context, siteId);
+    log.info(`Calculated cpc value: ${cpcValue} for site: ${siteId}`);
+    const projectedTrafficValue = projectedTrafficLost * cpcValue;
+
+    // Skip updating projected traffic data if lost traffic value is insignificant
+    return projectedTrafficValue > PROJECTED_VALUE_THRESHOLD
+      ? { projectedTrafficLost, projectedTrafficValue } : {};
+  } catch (err) {
+    log.warn(`Error while calculating projected traffic for ${auditUrl} : ${siteId}`, err);
+    return {};
+  }
 }
 
-export async function auditMetaTagsRunner(baseURL, context, site) {
+export async function auditMetaTagsRunner(auditUrl, context, site) {
   const { log, s3Client, dataAccess } = context;
   // Get top pages for a site
   const siteId = site.getId();
@@ -210,10 +229,16 @@ export async function auditMetaTagsRunner(baseURL, context, site) {
   const detectedTags = seoChecks.getDetectedTags();
 
   // Calculate projected traffic lost
-  const projectedTrafficLost = await calculateProjectedTraffic(context, site, detectedTags, log);
-  const cpcValue = await calculateCPCValue(context, site.getId());
-  log.info(`Calculated cpc value: ${cpcValue} for site: ${site.getId()}`);
-  const projectedTrafficValue = projectedTrafficLost * cpcValue;
+  const {
+    projectedTrafficLost,
+    projectedTrafficValue,
+  } = await calculateProjectedTraffic(
+    context,
+    auditUrl,
+    siteId,
+    detectedTags,
+    log,
+  );
 
   // Generate AI suggestions for detected tags if auto-suggest enabled for site
   const allTags = {
@@ -227,18 +252,18 @@ export async function auditMetaTagsRunner(baseURL, context, site) {
     detectedTags: updatedDetectedTags,
     sourceS3Folder: `${bucketName}/${prefix}`,
     fullAuditRef: '',
-    finalUrl: baseURL,
-    projectedTrafficLost,
-    projectedTrafficValue,
+    finalUrl: auditUrl,
+    ...(projectedTrafficLost && { projectedTrafficLost }),
+    ...(projectedTrafficValue && { projectedTrafficValue }),
   };
   return {
     auditResult,
-    fullAuditRef: baseURL,
+    fullAuditRef: auditUrl,
   };
 }
 
 export default new AuditBuilder()
-  .withUrlResolver(noopUrlResolver)
+  .withUrlResolver(wwwUrlResolver)
   .withRunner(auditMetaTagsRunner)
   .withPostProcessors([opportunityAndSuggestions])
   .build();