+++ test/audits/metatags.test.js
@@ -739,18 +739,10 @@ describe('Meta Tags', () => {
       expect(logStub.error).to.have.been.calledWith('Failed to extract tags from scraped content for bucket test-bucket and prefix scrapes/site-id/');
     });
 
-    it('should calculate projected traffic for detected tags', async () => {
+    it('should handle gracefully if Rum throws error', async () => {
       const RUMAPIClientStub = {
         createFrom: sinon.stub().returns({
-          query: sinon.stub().resolves([
-            {
-              url: 'http://example.com/blog/page1',
-              total: 100,
-              earned: 20,
-              owned: 70,
-              paid: 10,
-            },
-          ]),
+          query: sinon.stub().throws(new Error('RUM not available')),
         }),
       };
       const mockGetRUMDomainkey = sinon.stub().resolves('mockedDomainKey');
@@ -838,8 +830,6 @@ describe('Meta Tags', () => {
       await auditInstance.runner('http://example.com', context, site);
       expect(addAuditStub.calledWithMatch({
         auditResult: {
-          projectedTraffic: 200,
-          projectedTrafficValue: 200,
           detectedTags: {
             '/blog/page1': {
               h1: {
@@ -894,6 +884,334 @@ describe('Meta Tags', () => {
           },
         },
       }));
+    });
+
+    it('should skip updating projected traffic if value less than 500', async () => {
+      const RUMAPIClientStub = {
+        createFrom: sinon.stub().returns({
+          query: sinon.stub().resolves([
+            {
+              url: 'http://example.com/blog/page1',
+              total: 100,
+              earned: 20,
+              owned: 70,
+              paid: 10,
+            },
+          ]),
+        }),
+      };
+      const mockGetRUMDomainkey = sinon.stub().resolves('mockedDomainKey');
+      const metatagsOppty = {
+        getId: () => 'opportunity-id',
+        setAuditId: sinon.stub(),
+        save: sinon.stub(),
+        getSuggestions: sinon.stub().returns([]),
+        addSuggestions: sinon.stub().returns({ errorItems: [], createdItems: [1, 2, 3] }),
+        getType: () => 'meta-tags',
+      };
+      const site = {
+        getIsLive: sinon.stub().returns(true),
+        getId: sinon.stub().returns('site-id'),
+        getBaseURL: sinon.stub().returns('http://example.com'),
+      };
+      const topPages = [{ getURL: 'http://example.com/blog/page1', getTopKeyword: sinon.stub().returns('page') },
+        { getURL: 'http://example.com/blog/page2', getTopKeyword: sinon.stub().returns('Test') }];
+
+      dataAccessStub.Site.findById.resolves(site);
+      dataAccessStub.Configuration.findLatest.resolves({
+        isHandlerEnabledForSite: sinon.stub().returns(true),
+      });
+      dataAccessStub.SiteTopPage.allBySiteId.resolves(topPages);
+      dataAccessStub.Opportunity = {
+        allBySiteIdAndStatus: sinon.stub().returns([metatagsOppty]),
+      };
+      s3ClientStub.send
+        .withArgs(sinon.match.instanceOf(ListObjectsV2Command).and(sinon.match.has('input', {
+          Bucket: 'test-bucket',
+          Prefix: 'scrapes/site-id/',
+          MaxKeys: 1000,
+        })))
+        .resolves({
+          Contents: [
+            { Key: 'scrapes/site-id/blog/page1/scrape.json' },
+            { Key: 'scrapes/site-id/blog/page2/scrape.json' },
+          ],
+        });
+
+      s3ClientStub.send
+        .withArgs(sinon.match.instanceOf(GetObjectCommand).and(sinon.match.has('input', {
+          Bucket: 'test-bucket',
+          Key: 'scrapes/site-id/blog/page1/scrape.json',
+        }))).returns({
+          Body: {
+            transformToString: () => JSON.stringify({
+              scrapeResult: {
+                tags: {
+                  title: 'Test Page',
+                  description: '',
+                },
+              },
+            }),
+          },
+          ContentType: 'application/json',
+        });
+      s3ClientStub.send
+        .withArgs(sinon.match.instanceOf(GetObjectCommand).and(sinon.match.has('input', {
+          Bucket: 'test-bucket',
+          Key: 'scrapes/site-id/blog/page2/scrape.json',
+        }))).returns({
+          Body: {
+            transformToString: () => JSON.stringify({
+              scrapeResult: {
+                tags: {
+                  title: 'Test Page',
+                  h1: [
+                    'This is a dummy H1 that is intentionally made to be overly lengthy from SEO perspective',
+                  ],
+                },
+              },
+            }),
+          },
+          ContentType: 'application/json',
+        });
+      const getTopPagesForSiteStub = [
+        { getUrl: () => 'http://example.com/blog/page1' },
+        {
+          getUrl: () => 'http://example.com/blog/page2',
+        },
+        {
+          getUrl: () => 'http://example.com/',
+        },
+      ];
+      dataAccessStub.SiteTopPage.allBySiteId.resolves(topPages);
+      dataAccessStub.SiteTopPage.allBySiteIdAndSourceAndGeo.resolves(getTopPagesForSiteStub);
+      const addAuditStub = sinon.stub().resolves({ getId: () => 'audit-id' });
+      dataAccessStub.Audit.create = addAuditStub;
+      const mockCalculateCPCValue = sinon.stub().resolves(2);
+      const auditStub = await esmock('../../src/metatags/handler.js', {
+        '../../src/support/utils.js': { getRUMDomainkey: mockGetRUMDomainkey, calculateCPCValue: mockCalculateCPCValue },
+        '@adobe/spacecat-shared-rum-api-client': RUMAPIClientStub,
+      });
+      const auditInstance = auditStub.default;
+      const response = await auditInstance.runner('http://example.com', context, site);
+      expect(response).to.deep.equal({
+        auditResult: {
+          detectedTags: {
+            '/blog/page1': {
+              h1: {
+                seoImpact: 'High',
+                issue: 'Missing H1',
+                issueDetails: 'H1 tag is missing',
+                seoRecommendation: 'Should be present',
+              },
+              title: {
+                tagContent: 'Test Page',
+                seoImpact: 'High',
+                issue: 'Duplicate Title',
+                issueDetails: '2 pages share same title',
+                seoRecommendation: 'Unique across pages',
+              },
+              description: {
+                seoImpact: 'High',
+                issue: 'Empty Description',
+                issueDetails: 'Description tag is empty',
+                seoRecommendation: '140-160 characters long',
+              },
+            },
+            '/blog/page2': {
+              description: {
+                seoImpact: 'High',
+                issue: 'Missing Description',
+                issueDetails: 'Description tag is missing',
+                seoRecommendation: 'Should be present',
+              },
+              title: {
+                tagContent: 'Test Page',
+                seoImpact: 'High',
+                issue: 'Duplicate Title',
+                issueDetails: '2 pages share same title',
+                seoRecommendation: 'Unique across pages',
+              },
+              h1: {
+                tagContent: 'This is a dummy H1 that is intentionally made to be overly lengthy from SEO perspective',
+                seoImpact: 'Moderate',
+                issue: 'H1 too long',
+                issueDetails: '17 chars over limit',
+                seoRecommendation: 'Below 70 characters',
+              },
+            },
+          },
+          finalUrl: 'http://example.com',
+          fullAuditRef: '',
+          sourceS3Folder: 'test-bucket/scrapes/site-id/',
+        },
+        fullAuditRef: 'http://example.com',
+      });
+    }).timeout(5000);
+
+    it('should calculate projected traffic for detected tags', async () => {
+      const RUMAPIClientStub = {
+        createFrom: sinon.stub().returns({
+          query: sinon.stub().resolves([
+            {
+              url: 'http://example.com/blog/page1',
+              total: 100,
+              earned: 50000,
+              owned: 70,
+              paid: 10000,
+            },
+          ]),
+        }),
+      };
+      const mockGetRUMDomainkey = sinon.stub().resolves('mockedDomainKey');
+      const metatagsOppty = {
+        getId: () => 'opportunity-id',
+        setAuditId: sinon.stub(),
+        save: sinon.stub(),
+        getSuggestions: sinon.stub().returns([]),
+        addSuggestions: sinon.stub().returns({ errorItems: [], createdItems: [1, 2, 3] }),
+        getType: () => 'meta-tags',
+      };
+      const site = {
+        getIsLive: sinon.stub().returns(true),
+        getId: sinon.stub().returns('site-id'),
+        getBaseURL: sinon.stub().returns('http://example.com'),
+      };
+      const topPages = [{ getURL: 'http://example.com/blog/page1', getTopKeyword: sinon.stub().returns('page') },
+        { getURL: 'http://example.com/blog/page2', getTopKeyword: sinon.stub().returns('Test') }];
+
+      dataAccessStub.Site.findById.resolves(site);
+      dataAccessStub.Configuration.findLatest.resolves({
+        isHandlerEnabledForSite: sinon.stub().returns(true),
+      });
+      dataAccessStub.SiteTopPage.allBySiteId.resolves(topPages);
+      dataAccessStub.Opportunity = {
+        allBySiteIdAndStatus: sinon.stub().returns([metatagsOppty]),
+      };
+      s3ClientStub.send
+        .withArgs(sinon.match.instanceOf(ListObjectsV2Command).and(sinon.match.has('input', {
+          Bucket: 'test-bucket',
+          Prefix: 'scrapes/site-id/',
+          MaxKeys: 1000,
+        })))
+        .resolves({
+          Contents: [
+            { Key: 'scrapes/site-id/blog/page1/scrape.json' },
+            { Key: 'scrapes/site-id/blog/page2/scrape.json' },
+          ],
+        });
+
+      s3ClientStub.send
+        .withArgs(sinon.match.instanceOf(GetObjectCommand).and(sinon.match.has('input', {
+          Bucket: 'test-bucket',
+          Key: 'scrapes/site-id/blog/page1/scrape.json',
+        }))).returns({
+          Body: {
+            transformToString: () => JSON.stringify({
+              scrapeResult: {
+                tags: {
+                  title: 'Test Page',
+                  description: '',
+                },
+              },
+            }),
+          },
+          ContentType: 'application/json',
+        });
+      s3ClientStub.send
+        .withArgs(sinon.match.instanceOf(GetObjectCommand).and(sinon.match.has('input', {
+          Bucket: 'test-bucket',
+          Key: 'scrapes/site-id/blog/page2/scrape.json',
+        }))).returns({
+          Body: {
+            transformToString: () => JSON.stringify({
+              scrapeResult: {
+                tags: {
+                  title: 'Test Page',
+                  h1: [
+                    'This is a dummy H1 that is intentionally made to be overly lengthy from SEO perspective',
+                  ],
+                },
+              },
+            }),
+          },
+          ContentType: 'application/json',
+        });
+      const getTopPagesForSiteStub = [
+        { getUrl: () => 'http://example.com/blog/page1' },
+        {
+          getUrl: () => 'http://example.com/blog/page2',
+        },
+        {
+          getUrl: () => 'http://example.com/',
+        },
+      ];
+      dataAccessStub.SiteTopPage.allBySiteId.resolves(topPages);
+      dataAccessStub.SiteTopPage.allBySiteIdAndSourceAndGeo.resolves(getTopPagesForSiteStub);
+      const addAuditStub = sinon.stub().resolves({ getId: () => 'audit-id' });
+      dataAccessStub.Audit.create = addAuditStub;
+      const mockCalculateCPCValue = sinon.stub().resolves(2);
+      const auditStub = await esmock('../../src/metatags/handler.js', {
+        '../../src/support/utils.js': { getRUMDomainkey: mockGetRUMDomainkey, calculateCPCValue: mockCalculateCPCValue },
+        '@adobe/spacecat-shared-rum-api-client': RUMAPIClientStub,
+      });
+      const auditInstance = auditStub.default;
+      const response = await auditInstance.runner('http://example.com', context, site);
+      expect(response).to.deep.equal({
+        auditResult: {
+          projectedTrafficValue: 2400,
+          projectedTrafficLost: 1200,
+          detectedTags: {
+            '/blog/page1': {
+              h1: {
+                seoImpact: 'High',
+                issue: 'Missing H1',
+                issueDetails: 'H1 tag is missing',
+                seoRecommendation: 'Should be present',
+              },
+              title: {
+                tagContent: 'Test Page',
+                seoImpact: 'High',
+                issue: 'Duplicate Title',
+                issueDetails: '2 pages share same title',
+                seoRecommendation: 'Unique across pages',
+              },
+              description: {
+                seoImpact: 'High',
+                issue: 'Empty Description',
+                issueDetails: 'Description tag is empty',
+                seoRecommendation: '140-160 characters long',
+              },
+            },
+            '/blog/page2': {
+              description: {
+                seoImpact: 'High',
+                issue: 'Missing Description',
+                issueDetails: 'Description tag is missing',
+                seoRecommendation: 'Should be present',
+              },
+              title: {
+                tagContent: 'Test Page',
+                seoImpact: 'High',
+                issue: 'Duplicate Title',
+                issueDetails: '2 pages share same title',
+                seoRecommendation: 'Unique across pages',
+              },
+              h1: {
+                tagContent: 'This is a dummy H1 that is intentionally made to be overly lengthy from SEO perspective',
+                seoImpact: 'Moderate',
+                issue: 'H1 too long',
+                issueDetails: '17 chars over limit',
+                seoRecommendation: 'Below 70 characters',
+              },
+            },
+          },
+          finalUrl: 'http://example.com',
+          fullAuditRef: '',
+          sourceS3Folder: 'test-bucket/scrapes/site-id/',
+        },
+        fullAuditRef: 'http://example.com',
+      });
     }).timeout(5000);
   });