+++ test/fixtures/high-form-views-low-conversions.js
@@ -334,6 +334,21 @@ const testData = {
     auditResult: {
       formVitals: [
         {
+          url: 'https://www.surest.com/search',
+          formsubmit: {
+            'desktop:windows': 100,
+          },
+          formview: {},
+          formengagement: {
+            'desktop:windows': 700,
+            'mobile:ios': 300,
+          },
+          pageview: {
+            'desktop:windows': 5690,
+            'mobile:ios': 1000,
+          },
+        },
+        {
           url: 'https://www.surest.com/contact-us',
           formsubmit: {
             'desktop:windows': 100,
@@ -474,7 +489,7 @@ const testData = {
     runbook: 'https://adobe.sharepoint.com/:w:/s/AEM_Forms/EU_cqrV92jNIlz8q9gxGaOMBSRbcwT9FPpQX84bRKQ9Phw?e=Nw9ZRz',
     type: 'high-form-views-low-conversions',
     origin: 'AUTOMATION',
-    title: 'Form has high views but low conversions',
+    title: 'Form has low conversions',
     description: 'Form has high views but low conversions',
     tags: [
       'Forms Conversion',
@@ -495,6 +510,121 @@ const testData = {
           value: {
             page: 0.014947683109118086,
           },
+        },
+      ],
+    },
+    guidance: {},
+  },
+  opportunityData2: {
+    siteId: 'site-id',
+    auditId: 'audit-id',
+    runbook: 'https://adobe.sharepoint.com/:w:/s/AEM_Forms/EU_cqrV92jNIlz8q9gxGaOMBSRbcwT9FPpQX84bRKQ9Phw?e=Nw9ZRz',
+    type: 'high-form-views-low-conversions',
+    origin: 'AUTOMATION',
+    title: 'Form has low conversions',
+    description: 'Form has high views but low conversions',
+    tags: [
+      'Forms Conversion',
+    ],
+    data: {
+      form: 'https://www.surest.com/contact-us',
+      screenshot: '',
+      trackedFormKPIName: 'Conversion Rate',
+      trackedFormKPIValue: 0.014947683109118086,
+      formViews: 6690,
+      pageViews: 6690,
+      samples: 6690,
+      scrapedStatus: false,
+      metrics: [
+        {
+          type: 'conversionRate',
+          device: '*',
+          value: {
+            page: 0.014947683109118086,
+          },
+        },
+      ],
+    },
+    guidance: {},
+  },
+  opportunityData3: {
+    siteId: 'site-id',
+    auditId: 'audit-id',
+    runbook: 'https://adobe.sharepoint.com/:w:/s/AEM_Forms/EU_cqrV92jNIlz8q9gxGaOMBSRbcwT9FPpQX84bRKQ9Phw?e=Nw9ZRz',
+    type: 'high-form-views-low-conversions',
+    origin: 'AUTOMATION',
+    title: 'Form has low conversions',
+    description: 'Form has high views but low conversions',
+    tags: [
+      'Forms Conversion',
+    ],
+    data: {
+      form: 'https://www.surest.com/contact-us',
+      screenshot: '',
+      trackedFormKPIName: 'Conversion Rate',
+      trackedFormKPIValue: 0.014947683109118086,
+      formViews: 6690,
+      pageViews: 6690,
+      samples: 6690,
+      scrapedStatus: false,
+      metrics: [
+        {
+          type: 'conversionRate',
+          device: '*',
+          value: {
+            page: 0.014947683109118086,
+          },
+        },
+      ],
+    },
+    guidance: {
+      recommendations: [
+        {
+          insight: 'The form contains a large number of fields, which can be overwhelming and increase cognitive load for users',
+          recommendation: 'Consider using progressive disclosure techniques, such as multi-step forms, to make the process less daunting.',
+          type: 'guidance',
+          rationale: 'Progressive disclosure can help by breaking the form into smaller, more manageable steps that can decrease cognitive load and make it more likely for users to complete the form.',
+        },
+      ],
+    },
+  },
+  opportunityData4: {
+    siteId: 'site-id',
+    auditId: 'audit-id',
+    runbook: 'https://adobe.sharepoint.com/:w:/s/AEM_Forms/EU_cqrV92jNIlz8q9gxGaOMBSRbcwT9FPpQX84bRKQ9Phw?e=Nw9ZRz',
+    type: 'high-form-views-low-conversions',
+    origin: 'AUTOMATION',
+    title: 'Form has low conversions',
+    description: 'Form has high views but low conversions',
+    tags: [
+      'Forms Conversion',
+    ],
+    data: {
+      form: 'https://www.surest.com/contact-us',
+      screenshot: '',
+      trackedFormKPIName: 'Conversion Rate',
+      trackedFormKPIValue: 0.014947683109118086,
+      formViews: 6690,
+      pageViews: 6690,
+      samples: 6690,
+      scrapedStatus: true,
+      metrics: [
+        {
+          type: 'conversionRate',
+          device: '*',
+          value: {
+            page: 0.014947683109118086,
+          },
+        },
+      ],
+    },
+    guidance: {
+      recommendations: [
+        {
+          insight: 'The form has a conversion rate of 1%',
+          recommendation: 'Ensure that the form communicates a compelling reason for users to fill it out. ',
+          type: 'guidance',
+          rationale: 'A strong, benefit-driven headline and a concise supporting message can improve engagement.',
         },
       ],
     },